# Fix thread local use in AllocMonitor