--- a/PerfTools/AllocMonitorPreload/src/memory_proxies.cc
+++ b/PerfTools/AllocMonitorPreload/src/memory_proxies.cc
@@ -1,8 +1,18 @@
 #include <memory>
 #include <cassert>
 #include <atomic>
+#include <array>
+#include <mutex>
+#include <limits>
 #include <cstddef>
 #include <malloc.h>
+#define ALLOC_USE_PTHREADS
+#if defined(ALLOC_USE_PTHREADS)
+#include <pthread.h>
+#else
+#include <unistd.h>
+#include <sys/syscall.h>
+#endif
 
 #include "PerfTools/AllocMonitor/interface/AllocMonitorRegistry.h"
 #include "FWCore/Utilities/interface/thread_safety_macros.h"
@@ -30,6 +40,14 @@ namespace {
     return reinterpret_cast<T>(original);
   }
 
+  inline auto thread_id() {
+#if defined(ALLOC_USE_PTHREADS)
+    /*NOTE: if use pthread_self, the values returned by linux had                                                                                                                                                                                      lots of hash collisions when using a simple % hash. Worked                                                                                                                                                                                        better if first divided value by 0x700 and then did %.                                                                                                                                                                                            [test done on el8] */
+    return pthread_self();
+#else
+    return syscall(SYS_gettid);
+#endif
+  }
 #ifdef USE_LOCAL_MALLOC
   // this is a very simple-minded allocator used for any allocations
   // before we've finished our setup.  In particular, this avoids a
@@ -81,6 +99,60 @@ namespace {
 #else
   constexpr inline bool is_local_alloc(void* ptr) noexcept { return false; }
 #endif
+
+  struct ThreadTracker {
+    static constexpr unsigned int kEntries = 128;
+    using entry_type = decltype(thread_id());
+    std::array<std::atomic<entry_type>, kEntries> used_threads_;
+    std::array<std::mutex, kEntries> used_threads_mutex_;
+
+    ThreadTracker() {
+      //put a value which will not match the % used when looking up the entry
+      entry_type entry = 0;
+      for (auto& v : used_threads_) {
+        v = ++entry;
+      }
+    }
+
+    std::size_t thread_index(entry_type id) const {
+#if defined(ALLOC_USE_PTHREADS)
+      return (id / 0x700) % kEntries;
+#else
+      return id % kEntries;
+#endif
+    }
+
+    //returns true if the thread had not already stopped reporting
+    bool stop_reporting() {
+      auto id = thread_id();
+      auto index = thread_index(id);
+      //are we already in this thread?
+      if (id == used_threads_[index]) {
+        return false;
+      }
+      used_threads_mutex_[index].lock();
+      used_threads_[index] = id;
+      return true;
+    }
+
+    void start_reporting() {
+      auto id = thread_id();
+      auto index = thread_index(id);
+      auto& v = used_threads_[index];
+      if (v == static_cast<entry_type>(index + 1)) {
+        return;
+      }
+      assert(v == id);
+      v = index + 1;
+      used_threads_mutex_[index].unlock();
+    }
+  };
+
+  static ThreadTracker& getTracker() {
+    static ThreadTracker s_tracker;
+    return s_tracker;
+  }
+
 }  // namespace
 
 using namespace cms::perftools;
@@ -88,6 +160,10 @@ using namespace cms::perftools;
 extern "C" {
 void alloc_monitor_start() { alloc_monitor_running_state() = true; }
 void alloc_monitor_stop() { alloc_monitor_running_state() = false; }
+
+bool alloc_monitor_stop_thread_reporting() { return getTracker().stop_reporting(); }
+
+void alloc_monitor_start_thread_reporting() { getTracker().start_reporting(); }
 
 //----------------------------------------------------------------
 //C memory functions