#### PR description:

OS change in alma 8 now trigger calls to memory allocation during usage of a thread_local during dynamic loads. This update removes the use of thread_local and uses a hash based pre-set container to keep track of which threads have calls into the alloc routines.

#### PR validation:

Cod compilers and related unit test pass. Ran the MaxMemoryPreload tool using the first step of RelVal 8.0 which was failing without this change and runs file with this change.