@cmsbuild 2024-06-05
cms-bot internal usage<!-- bot cache: {"commits":{"cae2de49cf0dab1730d2f2753033b28f057f7d37":{"files":["PerfTools/AllocMonitor/interface/AllocMonitorRegistry.h","PerfTools/AllocMonitor/src/AllocMonitorRegistry.cc","PerfTools/AllocMonitorPreload/src/memory_proxies.cc"],"squashed":false,"time":1717608597}},"emoji":{},"last_seen_sha":"cae2de49cf0dab1730d2f2753033b28f057f7d37","signatures":{}} -->