@cmsbuild 2024-06-05
A new Pull Request was created by  @Dr15Jones for master.

It involves the following packages:

- PerfTools/AllocMonitor (**core**)
- PerfTools/AllocMonitorPreload (**core**)


@smuzaffar, @makortel, @Dr15Jones can you please review it and eventually sign? Thanks.
@felicepantaleo, @makortel this is something you requested to watch as well.
@antoniovilela, @rappoccio, @sextonkennedy you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>