@Dr15Jones 2024-06-05
I tried 4 different changes to AllocMonitor to work around the problem

1. This PR: Replacing the use of thread_local in AllocMonitor
2. keep track if a thread is attempting to read from the thread_local used in AllocMonitor
   1. Simple: only allow one thread at a time to attempt to read from the thread_local
   2. Complex: use a fixed size data structure which kept track of which threads were asking for the thread_local where the thread id was stored in the first available slot in the data structure
   3. Hash: use a fixed size data structure which kept track of which threads were asking for the thread_local where the thread id was used to for a hash into the container. If there was a hash collision, the colliding thread had to wait until that entry was available again.

The choice of which to use was based on a performance test where I ran 12 concurrent threads where each thread continuously new/delete an array of ints of size 5600 entries until the sum of all new/delete exceeded 100M.

| alloc monitor code used | total time of test (s) |
|-------------------------|---------------------|
| no monitoring | 11 s |
| pre PR | 65 s |
| this PR | 69 s |
| Simple | 165 s |
| Complex | 450 s |
| Hash | 70 s |