@Dr15Jones 2024-06-06
> looks like PR relvals are now taking over 4 hours ( and then killed by bot) if this change is used with --maxmem_profile flags

Yeah. Seems my performance test program was not a good representation of what cmsRun is doing as that test showed the old code and the new code having equivalent performance.