@Dr15Jones 2024-06-06
So using step3 of RelVal 8.0 (which is just what I had already on hand) then running 10 events and then taking the time between the start time of event 2 (to avoid variance from loading conditions from event 1) to the start time of event 10 (since that was the last timestamp in the job) I saw the following timings

| setup | time |
|-------|-----|
| no monitoring | 0.9s |
|original allocMonitor on old OS | 1.07s |
|this PR on old OS | 1.778s |
|this PR on new OS | 1.793s |
|updated PR on new OS |  1.044s |

The improvement came from me switching from a `syscall` to get the thread id to just `pthread_self` where the latter appears to just do some assembly call to get info from a register (which seems to be what the internal thread local of the OS is doing as well).