@smuzaffar 2024-06-06
test parameters:
- container = cmssw/el8:x86_64-d20240513